+++ build.sh
@@ -3,30 +3,30 @@
 set -e
 
 # Set version info
-export BOX_VERSION_BASE="1.5.0"
+export BOX_VERSION_BASE="1.6.0"
 
 # Set versions requested of main components (These will be used in Packer and passed to Ansible downstream)
 # The apt package versions can be determined with "apt-get madison <package name>", use the exact version number!
-export BOX_BASE="ilionx/ubuntu2204"
-export BOX_BASE_VERSION="1.5.0-20240320"
-export MINIKUBE_VERSION="1.32.0"
-export CRIDOCKERD_VERSION="0.3.11"
-export CRITOOLS_VERSION="1.28.0-1.1"
-export DOCKER_VERSION="5:25.0.5-1~ubuntu.22.04~jammy"
-export KUBE_REPO_VERSION="1.28"
-export KUBECTL_VERSION="1.28.3-1.1"
-export CNI_PLUGINS_VERSION="1.4.1"
-export HELM_VERSION="3.14.3"
+export BOX_BASE="ilionxde/ubuntu2204"
+export BOX_BASE_VERSION="1.6.0-20250321"
+export MINIKUBE_VERSION="1.35.0"
+export CRIDOCKERD_VERSION="0.3.16"
+export CRITOOLS_VERSION="1.32.0-1.1"
+export DOCKER_VERSION="5:28.0.2-1~ubuntu.22.04~jammy"
+export KUBE_REPO_VERSION="1.32"
+export KUBECTL_VERSION="1.32.3-1.1"
+export CNI_PLUGINS_VERSION="1.6.2"
+export HELM_VERSION="3.17.2"
 export KUBETAIL_VERSION="1.6.20"
 export KUBELOGS_VERSION="0.0.1"
 
 # Set versions of supported tools, if they don't match, a warning will be shown on screen
-export VIRTUALBOX_VERSION="7.0.14r161095"
-export PACKER_VERSION="1.10.2-dev"
-export VAGRANT_VERSION="2.4.1"
+export VIRTUALBOX_VERSION="7.1.6r167084"
+export PACKER_VERSION="v1.12.0"
+export VAGRANT_VERSION="2.4.3"
 
 # Set the Vagrant cloud user and box name (make sure you have admin permissions to, or are the owner of this repository)
-export VAGRANT_CLOUD_BOX_USER="ilionx"
+export VAGRANT_CLOUD_BOX_USER="ilionxde"
 export VAGRANT_CLOUD_BOX_NAME="ubuntu2204-minikube"
 
 # ############################################################################################## #
@@ -44,7 +44,7 @@ fi
 
 # Check the tool versions
 INSTALLED_VIRTUALBOX_VERSION=$(vboxmanage --version)
-INSTALLED_PACKER_VERSION=$(packer --version)
+INSTALLED_PACKER_VERSION=$(packer --version | awk '{print $2}')
 INSTALLED_VAGRANT_VERSION=$(vagrant --version | awk '{print $2}')
 
 if [[ "$INSTALLED_VIRTUALBOX_VERSION" != "$VIRTUALBOX_VERSION" || "$INSTALLED_PACKER_VERSION" != "$PACKER_VERSION" || "$INSTALLED_VAGRANT_VERSION" != "$VAGRANT_VERSION" ]]
@@ -66,25 +66,25 @@ then
     source build.env
 fi
 
-# Check if the variables VAGRANT_CLOUD_USER and VAGRANT_CLOUD_TOKEN have been set, if not ask for them
-if [ -z "$DEFAULT_VAGRANT_CLOUD_USER" ] || [ -z "$DEFAULT_VAGRANT_CLOUD_TOKEN" ]
+# Check if the variables HCP_CLIENT_ID and HCP_CLIENT_SECRET have been set, if not ask for them
+if [ -z "$DEFAULT_HCP_CLIENT_ID" ] || [ -z "$DEFAULT_HCP_CLIENT_SECRET" ]
 then
     # Ask user for vagrant cloud token
-    echo -n "What is your Vagrant Cloud username? [ilionx] "
+    echo -n "What is your HCP Service Principal Client ID? "
     read -r user
-    user=${user:-ilionx}
-    export VAGRANT_CLOUD_USER=${user}
+    echo ""
+    export HCP_CLIENT_ID=${user}
 
     # Ask user for vagrant cloud token
-    echo -n "What is your Vagrant Cloud token? "
+    echo -n "What is your HCP Service Principal Client Secret? "
     read -rs token
     echo ""
-    export VAGRANT_CLOUD_TOKEN=${token}
+    export HCP_CLIENT_SECRET=${token}
 else
-    export VAGRANT_CLOUD_USER=$DEFAULT_VAGRANT_CLOUD_USER
-    export VAGRANT_CLOUD_TOKEN=$DEFAULT_VAGRANT_CLOUD_TOKEN
+    export HCP_CLIENT_ID=$DEFAULT_HCP_CLIENT_ID
+    export HCP_CLIENT_SECRET=$DEFAULT_HCP_CLIENT_SECRET
 
-    echo "Your vagrant cloud user and token have been sourced from file build.env"
+    echo "Your HCP Client ID and Secret have been sourced from file build.env"
 fi
 
 # Export dynamic versioning info